+++ test/bigtest/interactors/manual-blocks.js
@@ -3,6 +3,7 @@ import {
   clickable,
   text,
   isPresent,
+  scoped,
 } from '@bigtest/interactor';
 
 import TestAreaInteractor from '@folio/stripes-components/lib/TextArea/tests/interactor'; // eslint-disable-line
@@ -14,6 +15,7 @@ import TextFieldInteractor from '@folio/stripes-components/lib/TextField/tests/i
 
 @interactor class PatronBlocksSection {
   sectionIsPresent = isPresent('#accordion-toggle-button-patronBlocksSection');
+  collapsePatronBlocks = clickable('#accordion-toggle-button-patronBlocksSection');
   patronBlocksAreVisible = isPresent('#patron-block-mcl');
   formIsVisible = isPresent('#patron-block-form');
 
@@ -29,7 +31,7 @@ import TextFieldInteractor from '@folio/stripes-components/lib/TextField/tests/i
     return this.when(() => this.formIsVisible);
   }
 
-  label = text('#accordion-toggle-button-patronBlocksSection > span > div > div > div:nth-child(1) > h3');
+  label = text('#accordion-toggle-button-patronBlocksSection [data-test-headline]');
 
   clickOnPatronBlockSection = clickable('#accordion-toggle-button-patronBlocksSection');
 
@@ -70,7 +72,7 @@ import TextFieldInteractor from '@folio/stripes-components/lib/TextField/tests/i
   title = text('#paneHeadertitle-patron-block-pane-title > h2 > span');
 
   // view
-  PatronBlockMessage = new TextFieldInteractor('#patron-block-place');
+  patronBlockMessage = scoped('#patron-block-place', TextFieldInteractor);
 }
 
 export default new PatronBlocksSection(5000);